test(package): update import paths
